support for unidirectional channels. message now piggybacks a return address.

+++ workspace/DistBank/src/bank/BankClient.java
@@ -5,12 +5,11 @@ import java.net.SocketAddress;
 import core.network.client.Client;
 
 import bank.messages.BankResponse;
-import bank.messages.ChangeRequest;
 import bank.messages.QueryRequest;
 
 public class BankClient extends Client {
 
-	public static int query(SocketAddress a) {
+	public static double query(SocketAddress a) {
 		QueryRequest req = new QueryRequest();
 		BankResponse resp = BankClient.exec(a, req);
 		return resp.getAmt();
